Create PayrollSystem.c
This is a Payroll System made by C Programming. Topics Covered is Arrays, Structs and File Handling. The system allows you to add employees, make pay slips according to the number of days present and absent due to festive or some other reasons, all the data is stored in a .dat file for the ease of checking.